Car Dealership
Developed a web site using HTML,javascript,AJAX,jquery for client side
scripting, php as server side scripting and mysql
• The web site was to implement a car dealership company to highlight
the interaction between customer and the car dealer and also maintain
the history of the orders placed
• HTML and CSS were used to design the user interface and javascript was
used to ensure the validations of registration pages. PHP was used to
retrieve data from mysql database
• The goal of the web site was to maintain the various cars and
accessories available where the user would be able to search for a
particular car or accessory with appropriate filters. The website also
enabled user to place orders and the user’s order history was maintained